+++ scripts/build.sh
@@ -526,14 +526,6 @@ case "$product-$platform-$method" in
       test
     ;;
 
-  VertexSample-*-*)
-    RunXcodebuild \
-      -project 'FirebaseVertexAI/Sample/VertexAISample.xcodeproj' \
-      -scheme "VertexAISample" \
-      "${xcb_flags[@]}" \
-      build
-    ;;
-
   Sessions-*-integration)
     # Perform "pod install" to install the relevant dependencies
     # ./FirebaseSessions/generate_testapp.sh