+++ FirebaseVertexAI/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 - [feature] The Vertex AI SDK no longer requires `@preconcurrency` when imported in Swift 6.
 - [feature] The Vertex AI Sample App now includes an image generation example.
+- [changed] The Vertex AI Sample App is now part of the
+  [quickstart-ios repo](https://github.com/firebase/quickstart-ios/tree/main/vertexai).
 
 # 11.9.0
 - [feature] **Public Preview**: Added support for generating images using the